Refactoring the Series1.rsc file into individual code prop files
Each of the code properties as described in the paper that we are
following now has its own Rascal src file in the 'codeProperties'
folder. I updated all imports to be no more than necessary but
including those that are, and every individual component works.

This is an initial refactoring. The Series1.rsc file should now
be extended to calculate the actual score. Moreover, the individual
files should probably be updated still to generate the exact output
we need.